+++ app/src/main/java/com/nikola/jakshic/dagger/vo/Bookmark.kt
@@ -7,5 +7,5 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "bookmark", indices = [(Index(value = ["account_id"], unique = true))])
 data class Bookmark(@ColumnInfo(name = "account_id") var id: Long) {
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "count") var count: Long = 0
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "count") var count: Int = 0
 }
