+++ app/src/main/java/com/nikola/jakshic/dagger/vo/Competitive.kt
@@ -10,11 +10,11 @@ import com.google.gson.annotations.SerializedName
 data class Competitive(
         @Expose @PrimaryKey @ColumnInfo(name = "match_id") @SerializedName("match_id") var matchId: Long,
         @Expose @ColumnInfo(name = "start_time") @SerializedName("start_time") var startTime: Long,
-        @Expose @ColumnInfo(name = "duration") @SerializedName("duration") var duration: Long,
+        @Expose @ColumnInfo(name = "duration") @SerializedName("duration") var duration: Int,
         @Expose @ColumnInfo(name = "radiant_name") @SerializedName("radiant_name") var radiantName: String?,
         @Expose @ColumnInfo(name = "dire_name") @SerializedName("dire_name") var direName: String?,
         @Expose @ColumnInfo(name = "league_name") @SerializedName("league_name") var leagueName: String?,
-        @Expose @ColumnInfo(name = "radiant_score") @SerializedName("radiant_score") var radiantScore: Long,
-        @Expose @ColumnInfo(name = "dire_score") @SerializedName("dire_score") var direScore: Long,
+        @Expose @ColumnInfo(name = "radiant_score") @SerializedName("radiant_score") var radiantScore: Int,
+        @Expose @ColumnInfo(name = "dire_score") @SerializedName("dire_score") var direScore: Int,
         @Expose @ColumnInfo(name = "radiant_win") @SerializedName("radiant_win") var isRadiantWin: Boolean
 )
