+++ app/src/main/java/com/nikola/jakshic/dagger/vo/Hero.kt
@@ -8,6 +8,6 @@ import com.google.gson.annotations.SerializedName
 @Entity(tableName = "heroes", primaryKeys = arrayOf("account_id", "hero_id"))
 data class Hero(
         @ColumnInfo(name = "account_id") var accountId: Long,
-        @ColumnInfo(name = "hero_id") @SerializedName("hero_id") @Expose var heroId: Long,
+        @ColumnInfo(name = "hero_id") @SerializedName("hero_id") @Expose var heroId: Int,
         @ColumnInfo(name = "games") @SerializedName("games") @Expose var gamesPlayed: Int,
         @ColumnInfo(name = "wins") @SerializedName("win") @Expose var gamesWon: Int)
