+++ app/src/main/java/com/nikola/jakshic/dagger/vo/SearchHistory.kt
@@ -8,5 +8,5 @@ import android.arch.persistence.room.PrimaryKey
 @Entity(tableName = "search_history", indices = [Index(value = ["query"], unique = true)])
 data class SearchHistory(
         @ColumnInfo(collate = ColumnInfo.NOCASE) var query: String) {
-    @PrimaryKey(autoGenerate = true) var id: Long = 0
+    @PrimaryKey(autoGenerate = true) var id: Int = 0
 }
